requirements.txt
asgiref==3.11.1
Django==6.0.2
django-taggit==6.1.0
gunicorn==23.0.0
Markdown==3.10.2
nh3==0.3.3
packaging==26.0
python-dotenv==1.2.1
sqlparse==0.5.5
unicode==2.9
Unidecode==1.4.0
whitenoise==6.11.0
